Move bearer token flag and static user authenticator into authoptions
