+++ pkg/auth/oauth2/auth_oidc_test.go
@@ -1,4 +1,4 @@
-package auth
+package oauth2
 
 import (
 	"context"
